g-i: only workaround #757818 on squeeze+wheezy, which are affected by the bug. this speeds up jessie tests.

+++ bin/g-i-installation.sh
@@ -130,8 +130,14 @@ bootstrap_system() {
 	cd $WORKSPACE
 	echo "Creating throw-away logical volume with ${DISKSIZE_IN_GB} GiB now."
 	sudo lvcreate -L${DISKSIZE_IN_GB}G -n $NAME $VG
-	echo "Workaround to remove swap signature from previous installs, see #757818"
-	time sudo dd if=/dev/zero of=$LV bs=4096 || true
+	# FIXME: the following will become unneeded / simpler once squeeze+wheezy are not tested anymore
+	RELEASE="$(echo $NAME | cut -d "_" -f2)"
+	if [ "$RELEASE" != "jessie" ] ; then
+		echo "Workaround to remove swap signature from previous installs, see #757818"
+		time sudo dd if=/dev/zero of=$LV bs=4096 || true
+	else
+		echo "Release $RELEASE detected, where #757818 should be fixed, thus not applying workaround..."
+	fi
 	echo "Creating raw disk image with ${DISKSIZE_IN_GB} GiB now."
 	sudo qemu-img create -f raw $LV ${DISKSIZE_IN_GB}G
 	echo "Doing g-i installation test for $NAME now."
